Move Gaussian algorithms into ops.gaussian
